different types of parsers and their performance update

+++ ingestion/schemas/invoice_schema.py
@@ -3,4 +3,4 @@ from pydantic import BaseModel
 class Invoice(BaseModel):
     invoice_number: str
     invoice_date: str
-    total_amount: int 
+    total_amount: float
